This closes #9 and #7.
The main APIS have been implemented and tested. Now using hypothesis
tests as well. The code has been improved to be more generic.
All tests are seperated and there is only a main readme file.

In the older version we were keeping a readme file
for each of the APIs. Those will now be replaced by the main
readme file.

Seperating tests and added Hypothesis.

Hypothesis tests have been added to the library.
Also each api now has an individual test file.

Additions and Corrections.

Several corrections has been made to the functions. Such as, create the url
correctly and checked if the arguments for each api are correct.

Also I have added some new functions. Now we can validate the post
and in order to so a lower case function was also added.

Added hypothesis.

Thus the requirements and the ignore files were altered.

The main readme file.