fix parameters.
Parameters where being passed in fix parameters function
as a dictionary. Fix parameters function is in charge of creating the url.

This was causing an issue because the user had to pass a dictionary
with all the parameters. This is now removed and the user can pass
an argument straight to the function. Example:

 paramaters = api.fix_parameters(title='sun')

To implement this change the passing arguments for the fix parameters
have change. Followed by alterations to each apis fix parameters function.

Tests have been re writen, tested and failed and then passed.