Preparation for version one (#57)
* Add to results set for arxiv

* revisit arxiv

Closes #41 and #47

Add new result fields for arxiv and implement new
search arguments.

Morever removes unittests and hypothesis from test_arxiv.

Partial #46. New section is being written for the documentation
for exploring each api individually.

* revisit plos

Closes #45 and #48

Add to the result fields, doi and url. Also new search arguments
journal and subject. Remove unittest and hypothesis from the tests.

Partial work on #46.

* revisit plos

Closes #44 and #50

Add to the result field doi and url. Also add search arguments
such as journals and category. Remove unittest and hypothesis.

Partial work on #46. Need to chat with Ger

* revisit springer

Closes #43 and #51

Add to the result set fields such as doi and url. Add to the
search fields subject and add message for fields that are not
implementable. Remove unittests and hypothesis which are
replaced with pytests.

Partial work on #46.

* revisit ieee

Closes #39 and #41

Ieee have updated their api and a key is needed now
to interact with it.

Removes unittests and hypothesis, move to pytest.

* remove validate function

No need for this validation function.
Could implement a new one but it would be different

* Closes #52

* Fix issue with tests on Travis

Remove requirements as they are listed on setup.py.
Also minor tweak for readability

* Closes #53

* add contributions.rst

* standarise results.

Now all APIs return the same columns, if a key is not available
by the API then the key value is set to 'Not available'

* add open access to results

* update tutorials

* fix reference sections

Search arguments and results

* update welcome

* Finish of Guides

* Finish off References

* minor additions and tweaks

* fix failure

* version 1.0.0

* language tweaks